feat(models-table): Implement headless table for models with sorting, filtering, and pagination
