+++ src/lib/stores/models.ts
@@ -1,0 +1,3 @@
+import { writable } from "svelte/store";
+
+export const defaultModelStore = writable<string>("");
