+++ src/routes/models/cards/ParameterSlider.svelte
